super_club_backend/pyproject.toml
[project]
name = "super-club-backend"
version = "0.1.0"
description = "Super Club API 后端服务（FastAPI + PostgreSQL + SQLAlchemy）"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "sqlalchemy==2.0.25",
    "pymysql==1.1.0",
    "alembic==1.13.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "email-validator>=2.3.0",
    "requests>=2.32.5",
]

[tool.uv]

[tool.uv.sources]
# 可在此添加私有源镜像，如需要：
# "pypi" = { url = "https://pypi.tuna.tsinghua.edu.cn/simple" }

[project.scripts]
# 统一启动命令
api = "uvicorn app.main:app --host 0.0.0.0 --port 8001 --reload"
initdb = "python init_db.py"


